dev(dashboard): search and home sections
