+++ src/features/dashboard/content.style.js
@@ -3,7 +3,7 @@ import { ToggleButton } from '@material-ui/lab'
 
 export default makeStyles((theme) => ({
   root: {
-    padding: theme.spacing(2, 0, 0, 2),
+    padding: theme.spacing(0, 2),
     marginLeft: 0,
     transition: theme.transitions.create('margin', {
       easing: theme.transitions.easing.sharp,
@@ -20,19 +20,5 @@ export default makeStyles((theme) => ({
     [theme.breakpoints.up('md')]: {
       marginLeft: 283.4
     }
-  },
-  filter_root: {
-    border: 'none'
-  }
-}))
-
-export const FilterButton = styled(ToggleButton)(({ theme }) => ({
-  '&.MuiToggleButton-root': {
-    border: 'none',
-    padding: theme.spacing(1)
-  },
-  '&.Mui-selected': {
-    color: theme.palette.primary.main,
-    backgroundColor: theme.palette.background.default
   }
 }))
