+++ src/features/dashboard/workspace.style.js
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core'
+
+export default makeStyles((theme) => ({
+  title: {
+    ...theme.typography.h5,
+    padding: theme.spacing(2, 0)
+  }
+}))
