+++ src/mocks/api/home.handler.js
@@ -1,5 +1,5 @@
 import data from '../data/courses.json'
 
 export function courses() {
-  return data
+  return data.slice(0, 10)
 }
